requirements.txt
alembic           1.16.4
blinker           1.9.0
click             8.2.1
colorama          0.4.6
dnspython         2.7.0
email_validator   2.2.0
Flask             3.1.1
Flask-Login       0.6.3
Flask-Migrate     4.1.0
Flask-Moment      1.0.6
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.2
greenlet          3.2.4
idna              3.10
itsdangerous      2.2.0
Jinja2            3.1.6
Mako              1.3.10
MarkupSafe        3.0.2
packaging         25.0
pip               25.1.1
python-dotenv     1.1.1
SQLAlchemy        2.0.42
typing_extensions 4.14.1
Werkzeug          3.1.3
WTForms           3.2.1
gunicorn==21.2.0